Add files via upload
This is a document which shows a theory relating how the looping of time can lead to unbounded possibilities. Although this may not be entirely correct, the concept can give a new perspective in studying cosmology. It is not a proper;y researched documentation, but a theory paper which lies in the realm of speculation.